+++ vector/src/fdroid/java/im/vector/app/fdroid/features/settings/troubleshoot/TestBatteryOptimization.kt
@@ -15,29 +15,30 @@
  */
 package im.vector.app.fdroid.features.settings.troubleshoot
 
-import androidx.fragment.app.Fragment
+import androidx.appcompat.app.AppCompatActivity
 import im.vector.app.R
+import im.vector.app.core.resources.StringProvider
 import im.vector.app.core.utils.isIgnoringBatteryOptimizations
 import im.vector.app.core.utils.requestDisablingBatteryOptimization
 import im.vector.app.features.settings.troubleshoot.NotificationTroubleshootTestManager
 import im.vector.app.features.settings.troubleshoot.TroubleshootTest
+import javax.inject.Inject
 
-// Not used anymore
-class TestBatteryOptimization(val fragment: Fragment) : TroubleshootTest(R.string.settings_troubleshoot_test_battery_title) {
+class TestBatteryOptimization @Inject constructor(
+        private val context: AppCompatActivity,
+        private val stringProvider: StringProvider
+) : TroubleshootTest(R.string.settings_troubleshoot_test_battery_title) {
 
     override fun perform() {
-        val context = fragment.context
-        if (context != null && isIgnoringBatteryOptimizations(context)) {
-            description = fragment.getString(R.string.settings_troubleshoot_test_battery_success)
+        if (isIgnoringBatteryOptimizations(context)) {
+            description = stringProvider.getString(R.string.settings_troubleshoot_test_battery_success)
             status = TestStatus.SUCCESS
             quickFix = null
         } else {
-            description = fragment.getString(R.string.settings_troubleshoot_test_battery_failed)
+            description = stringProvider.getString(R.string.settings_troubleshoot_test_battery_failed)
             quickFix = object : TroubleshootQuickFix(R.string.settings_troubleshoot_test_battery_quickfix) {
                 override fun doFix() {
-                    fragment.activity?.let {
-                        requestDisablingBatteryOptimization(it, fragment, NotificationTroubleshootTestManager.REQ_CODE_FIX)
-                    }
+                    requestDisablingBatteryOptimization(context, null, NotificationTroubleshootTestManager.REQ_CODE_FIX)
                 }
             }
             status = TestStatus.FAILED